mcp-server/requirements.txt
# FastMCP
fastmcp>=0.4.1

# Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.6.1
python-dateutil>=2.8.0
